Migrations and new Readme

+++ src/repositories/implementations/postMysqlImplementation.php
@@ -1,6 +1,6 @@
 <?php
-require_once './src/repositories/IPostRepository.php';
-require_once './src/entities/post.php';
+require_once '../src/repositories/IPostRepository.php';
+require_once '../src/entities/post.php';
 
 
 class PostMysql implements IPost{
@@ -10,8 +10,8 @@ class PostMysql implements IPost{
     ){}
 
     function save(Post $post, $author_id){
+        if(!$this->connection) return throw new Exception('Error connecting to database');
         $sql = 'INSERT INTO Posts (author_id, author,  body, created_at, videoUrl) VALUES (:author_id, :author, :body, :date, :videoUrl);';
-        
         $query = $this->connection->prepare($sql);
         $query->bindValue(":author_id", $author_id);
         $query->bindValue(":author", $post->author);
@@ -24,6 +24,7 @@ class PostMysql implements IPost{
     }
 
     function get($author, $id){
+        if(!$this->connection) return throw new Exception('Error connecting to database');
         $sql = 'SELECT * FROM POSTS WHERE author = :author AND _id == :id';
         $data = $this->connection->prepare($sql);
         $data->bindValue(":author", $author);
@@ -37,6 +38,7 @@ class PostMysql implements IPost{
 
 
     function getAll($author){
+        if(!$this->connection) return throw new Exception('Error connecting to database');
         $sql = 'SELECT * FROM Posts WHERE author = :author';
         $data = $this->connection->prepare($sql);
         $data->bindValue(":author", $author);
@@ -48,6 +50,7 @@ class PostMysql implements IPost{
     }
 
     function delete($id, $username){
+        if(!$this->connection) return throw new Exception('Error connecting to database');
         $sql = 'DELETE FROM Posts Where id = :id AND author = :author';
         $executeDelete = $this->connection->prepare($sql);
         $executeDelete->bindValue(":id", $id);
